split up session into a separate file
