+++ js/genkit/src/environment.ts
@@ -14,13 +14,7 @@
  * limitations under the License.
  */
 
-import {
-  GenerateOptions,
-  MessageData,
-  PromptFn,
-  ToolAction,
-  ToolConfig,
-} from '@genkit-ai/ai';
+import { PromptFn, ToolAction, ToolConfig } from '@genkit-ai/ai';
 import {
   CallableFlow,
   FlowConfig,
@@ -30,14 +24,20 @@ import {
   z,
 } from '@genkit-ai/core';
 import { PromptMetadata } from '@genkit-ai/dotprompt';
-import { AsyncLocalStorage } from 'node:async_hooks';
 import { v4 as uuidv4 } from 'uuid';
-import { Chat, ChatOptions, MAIN_THREAD } from './chat';
-import { ExecutablePrompt, Genkit } from './genkit';
+import { MAIN_THREAD } from './chat.js';
+import { ExecutablePrompt, Genkit } from './genkit.js';
+import {
+  Session,
+  SessionData,
+  SessionError,
+  SessionOptions,
+  SessionStore,
+  getCurrentSession,
+  inMemorySessionStore,
+} from './session.js';
 
-export type BaseGenerateOptions = Omit<GenerateOptions, 'prompt'>;
-
-type EnvironmentType = Pick<
+type EnvironmentInterface = Pick<
   Genkit,
   'defineFlow' | 'defineStreamingFlow' | 'defineTool' | 'definePrompt'
 >;
@@ -63,7 +63,9 @@ type EnvironmentSessionOptions<S extends z.ZodTypeAny> = Omit<
  * await session.runFlow(flow, {...})
  * ```
  */
-export class Environment<S extends z.ZodTypeAny> implements EnvironmentType {
+export class Environment<S extends z.ZodTypeAny>
+  implements EnvironmentInterface
+{
   private store: SessionStore<S>;
 
   constructor(
@@ -260,202 +262,5 @@ export class Environment<S extends z.ZodTypeAny> implements EnvironmentType {
       throw new SessionError('not running within a session');
     }
     return currentSession as Session<S>;
-  }
-}
-
-export type SessionOptions<S extends z.ZodTypeAny> = BaseGenerateOptions & {
-  /** Schema describing the state. */
-  stateSchema?: S;
-  /** Session store implementation for persisting the session state. */
-  store?: SessionStore<S>;
-  /** Initial state of the session.  */
-  state?: z.infer<S>;
-  /** Custom session Id. */
-  sessionId?: string;
-};
-
-/**
- * Session encapsulates a statful execution environment for chat.
- * Chat session executed within a session in this environment will have acesss to
- * session session convesation history.
- *
- * ```ts
- * const ai = genkit({...});
- * const chat = ai.chat(); // create a Session
- * let response = await chat.send('hi'); // session/history aware conversation
- * response = await chat.send('tell me a story');
- * ```
- */
-export class Session<S extends z.ZodTypeAny> {
-  readonly id: string;
-  readonly schema?: S;
-  private sessionData?: SessionData<S>;
-  private store: SessionStore<S>;
-
-  constructor(
-    readonly parent: Genkit | Environment<S>,
-    private requestBase?: BaseGenerateOptions,
-    options?: {
-      id?: string;
-      stateSchema?: S;
-      sessionData?: SessionData<S>;
-      store?: SessionStore<S>;
-    }
-  ) {
-    this.id = options?.id ?? uuidv4();
-    this.schema = options?.stateSchema;
-    this.sessionData = options?.sessionData;
-    if (!this.sessionData) {
-      this.sessionData = {};
-    }
-    if (!this.sessionData.threads) {
-      this.sessionData!.threads = {};
-    }
-    this.store = options?.store ?? new InMemorySessionStore();
-  }
-
-  get genkit(): Genkit {
-    if (this.parent instanceof Session) {
-      return this.parent.genkit;
-    }
-    if (this.parent instanceof Environment) {
-      return this.parent.genkit;
-    }
-    return this.parent;
-  }
-
-  get state(): z.infer<S> {
-    // We always get state from the parent. Parent session is the source of truth.
-    if (this.parent instanceof Session) {
-      return this.parent.state;
-    }
-    return this.sessionData!.state;
-  }
-
-  async updateState(data: z.infer<S>): Promise<void> {
-    // We always update the state on the parent. Parent session is the source of truth.
-    if (this.parent instanceof Session) {
-      return this.parent.updateState(data);
-    }
-    let sessionData = await this.store.get(this.id);
-    if (!sessionData) {
-      sessionData = {} as SessionData<S>;
-    }
-    sessionData.state = data;
-    this.sessionData = sessionData;
-
-    await this.store.save(this.id, sessionData);
-  }
-
-  chat<S extends z.ZodTypeAny = z.ZodTypeAny>(
-    options?: ChatOptions<S>
-  ): Chat<S>;
-
-  chat<S extends z.ZodTypeAny = z.ZodTypeAny>(
-    threadName: string,
-    options?: ChatOptions<S>
-  ): Chat<S>;
-
-  /**
-   * Create a chat session with the provided options.
-   *
-   * ```ts
-   * const chat = ai.chat({
-   *   system: 'talk like a pirate',
-   * })
-   * let response = await chat.send('tell me a joke')
-   * response = await chat.send('another one')
-   * ```
-   */
-  chat(
-    optionsOrThreadName?: ChatOptions<S> | string,
-    maybeOptions?: ChatOptions<S>
-  ): Chat<S> {
-    let options: ChatOptions<S> | undefined;
-    let threadName = '__main';
-    if (maybeOptions) {
-      threadName = optionsOrThreadName as string;
-      options = maybeOptions as ChatOptions<S>;
-    } else if (optionsOrThreadName) {
-      if (typeof optionsOrThreadName === 'string') {
-        threadName = optionsOrThreadName as string;
-      } else {
-        options = optionsOrThreadName as ChatOptions<S>;
-      }
-    }
-    return new Chat<S>(
-      this,
-      {
-        ...this.requestBase,
-        ...options,
-      },
-      {
-        threadName,
-        id: this.id,
-        sessionData: {
-          state: options?.state,
-        },
-        stateSchema: options?.stateSchema,
-        store: this.store ?? options?.store,
-      }
-    );
-  }
-
-  toJSON() {
-    return this.sessionData;
-  }
-}
-
-export interface SessionData<S extends z.ZodTypeAny> {
-  state?: z.infer<S>;
-  threads?: Record<string, MessageData[]>;
-}
-
-const sessionAls = new AsyncLocalStorage<Session<any>>();
-
-/**
- * Executes provided function within the provided session state.
- */
-export function runWithSession<S extends z.ZodTypeAny, O>(
-  session: Session<S>,
-  fn: () => O
-): O {
-  return sessionAls.run(session, fn);
-}
-
-/** Returns the current session. */
-export function getCurrentSession<S extends z.ZodTypeAny = z.ZodTypeAny>():
-  | Session<S>
-  | undefined {
-  return sessionAls.getStore();
-}
-
-/** Throw when session state errors occur, ex. missing state, etc. */
-export class SessionError extends Error {
-  constructor(msg: string) {
-    super(msg);
-  }
-}
-
-/** Session store persists session data such as state and chat messages. */
-export interface SessionStore<S extends z.ZodTypeAny> {
-  get(sessionId: string): Promise<SessionData<S> | undefined>;
-
-  save(sessionId: string, data: SessionData<S>): Promise<void>;
-}
-
-export function inMemorySessionStore() {
-  return new InMemorySessionStore();
-}
-
-class InMemorySessionStore<S extends z.ZodTypeAny> implements SessionStore<S> {
-  private data: Record<string, SessionData<S>> = {};
-
-  async get(sessionId: string): Promise<SessionData<S> | undefined> {
-    return this.data[sessionId];
-  }
-
-  async save(sessionId: string, sessionData: SessionData<S>): Promise<void> {
-    this.data[sessionId] = sessionData;
   }
 }
